@@ -3,22 +3,20 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.10.2",
-    "@material-ui/icons": "^4.9.1",
-    "@material-ui/lab": "^4.0.0-alpha.56",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.5.0",
-    "@testing-library/user-event": "^7.2.1",
-    "chart.js": "^2.9.3",
-    "enzyme": "^3.11.0",
-    "enzyme-adapter-react-16": "^1.15.2",
-    "react": "^16.13.1",
-    "react-chartjs-2": "^2.9.0",
-    "react-dom": "^16.13.1",
-    "react-redux": "^7.2.0",
-    "react-scripts": "3.4.1",
-    "react-spinners": "^0.8.3",
-    "redux": "^4.0.5",
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.1",
+    "@mui/icons-material": "^7.3.2",
+    "@mui/material": "^7.3.2",
+    "@testing-library/react": "^16.3.0",
+    "@testing-library/user-event": "^14.6.1",
+    "chart.js": "^4.5.0",
+    "react": "^19.1.1",
+    "react-chartjs-2": "^5.3.0",
+    "react-dom": "^19.1.1",
+    "react-redux": "^9.2.0",
+    "react-scripts": "^5.0.1",
+    "react-spinners": "^0.17.0",
+    "redux": "^5.0.1",
     "redux-mock-store": "^1.5.4",
     "redux-saga": "^1.1.3",
     "redux-saga-tester": "^1.0.812"
@@ -45,6 +43,17 @@
     ]
   },
   "devDependencies": {
-    "jest-canvas-mock": "^2.2.0"
+    "@testing-library/jest-dom": "^6.8.0",
+    "@types/chart.js": "^2.9.41",
+    "@types/jest": "^30.0.0",
+    "@types/node": "^24.3.3",
+    "@types/react": "^19.1.13",
+    "@types/react-dom": "^19.1.9",
+    "@types/react-redux": "^7.1.34",
+    "@types/redux": "^3.6.31",
+    "@types/redux-mock-store": "^1.5.0",
+    "@types/redux-saga": "^0.9.31",
+    "jest-canvas-mock": "^2.2.0",
+    "typescript": "^5.9.2"
   }
 }
