@@ -1,17 +1,25 @@
-import React from "react";
+import React, { ReactNode, ErrorInfo } from "react";
 
-class ErrorBoundary extends React.Component {
-  constructor(props) {
+interface Props {
+  children: ReactNode;
+}
+
+interface State {
+  hasError: boolean;
+}
+
+class ErrorBoundary extends React.Component<Props, State> {
+  constructor(props: Props) {
     super(props);
     this.state = { hasError: false };
   }
 
-  static getDerivedStateFromError(error) {
+  static getDerivedStateFromError(error: Error): State {
     // Update state so the next render will show the fallback UI.
     return { hasError: true };
   }
 
-  componentDidCatch(error, errorInfo) {
+  componentDidCatch(error: Error, errorInfo: ErrorInfo) {
     // You can also log the error to an error reporting service
     // logErrorToMyService(error, errorInfo);
   }
@@ -26,7 +34,7 @@ class ErrorBoundary extends React.Component {
   }
 
   reset() {
-    this.state = { hasError: false };
+    this.setState({ hasError: false });
   }
 }
 
