@@ -1,12 +1,13 @@
 import * as types from "../constants/ActionTypes";
 import { periodOptions } from "../constants/DropdownOptions";
+import { FilterState, ActionType } from "../types";
 
-const defaultState = {
+const defaultState: FilterState = {
   symbol: "",
   period: periodOptions[0].key,
 };
 
-const filter = (state = defaultState, action) => {
+const filter = (state = defaultState, action: ActionType): FilterState => {
   switch (action.type) {
     case types.SYMBOL_SELECTED:
       return {
