@@ -1,11 +1,12 @@
 import * as types from "../constants/ActionTypes";
+import { SearchState, ActionType, DropdownOption } from "../types";
 
-function normalizeSearchResults(rawData) {
+function normalizeSearchResults(rawData: any): DropdownOption[] | {} {
   if (!(rawData && rawData.bestMatches)) {
     return {};
   }
-  const options = [];
-  rawData.bestMatches.forEach((bm) => {
+  const options: DropdownOption[] = [];
+  rawData.bestMatches.forEach((bm: any) => {
     options.push({
       value: bm["2. name"],
       key: bm["1. symbol"],
@@ -14,11 +15,11 @@ function normalizeSearchResults(rawData) {
   return options;
 }
 
-const defaultState = {
+const defaultState: SearchState = {
   options: [],
 };
 
-const search = (state = defaultState, action) => {
+const search = (state = defaultState, action: ActionType): SearchState => {
   switch (action.type) {
     case types.SYMBOL_FOUND:
       return {
