@@ -2,7 +2,7 @@
 // allows you to do things like:
 // expect(element).toHaveTextContent(/react/i)
 // learn more: https://github.com/testing-library/jest-dom
-import '@testing-library/jest-dom/extend-expect';
+import '@testing-library/jest-dom';
 import 'jest-canvas-mock';
 
 document.createRange = () => ({
@@ -12,4 +12,4 @@ document.createRange = () => ({
       nodeName: "BODY",
       ownerDocument: document,
     },
-  })
+  } as any);
